+++ CHANGES.markdown
@@ -1,3 +1,6 @@
+## Changes in TBD
+ - `Ord` instance for `TyCon` and `TypeRep`.
+
 ## Changes in 0.8.3 [2020.09.30]
  - Backport new instances from GHC 9.0/`base-4.15`
    * `MonadFix` and `MonadZip` instances for `Complex`